openfaas/generate_totp/requirements.txt
psycopg2-binary==2.9.3
cryptography==36.0.1
pyotp
pillow==9.0.0